Updated Utilities & Javander & User Classes
Now, Javander has an input method, and the String processId  moved to Utilities Class.

Added a real function to Register on Utilities class.

Now I have uncommented getGlobalId method and moved it to Utilities class, it gets data from a server(in this case of "http://httpbin.org/ip", a JSON example).