Verification Email
I separated the files by category, and moved some functions so that they fit better in the context of the program. Handing over the responsibility of several functions to the Process class.

I created the SendEmail class that sends a verification email to the user containing the process code, which will later be exchanged for the confirmation code.